# [SPARK-8314][MLlib] improvement in performance of MLUtils.appendBias