@dbtsai 2015-06-13
Merged into master. Thanks.